Implemented initial version of download daemon.
- The code is found in BackgroundJob.py.
- It starts when the server is run.
- The code for the background job found in BackgroundJob.py is commented since this daemon is not expected to be executed locally, but rather on a dedicated server, since it executes expensive operations.
- The API and files to execute the download operations were modified accordingly to the implementation of the daemon.

+++ sampleDownload/Load.py
@@ -2,7 +2,9 @@ import os
 from zipfile import ZipFile
 
 
-def load(zip_file, destination_dir):
+def load(zip_file: str, destination_dir: str) -> None:
+    if len(zip_file) == 0:
+        return
     delete_previous_sample(destination_dir)  # Delete the images for the previous sample
     unzip_current_sample(zip_file, destination_dir)  # Unzip and save the images for the current ZIP file
     os.remove(zip_file)  # Remove the downloaded ZIP file
@@ -23,7 +25,3 @@ def unzip_current_sample(zip_file, destination_dir):
     for f in file_list:
         if not f.endswith(".png"):
             os.remove(os.path.join(destination_dir, f))
-
-
-if __name__ == "__main__":
-    load("D20201211T213045_IFCB109.zip", "../production")
